% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplr.R
\name{peek_middle}
\alias{peek_middle}
\title{Get the middle n rows of the dataframe}
\usage{
peek_middle(dat, n = 10)
}
\arguments{
\item{dat}{Table to peek}

\item{n}{Number of rows to peek}
}
\description{
Get the middle n rows of the dataframe
}
