% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplr.R
\name{peek_random}
\alias{peek_random}
\title{Get n random rows of the dataframe sorted by row index}
\usage{
peek_random(dat, n = 10)
}
\arguments{
\item{dat}{Dataframe for row peeking}

\item{n}{Number of random rows to peek}
}
\value{
Dataframe of peeked elements
}
\description{
Get n random rows of the dataframe sorted by row index
}
