% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{skim_formatted}
\alias{skim_formatted}
\title{Get overview over the data using the skimr packages}
\usage{
skim_formatted(dat, digits = 4)
}
\arguments{
\item{dat}{Dataframe for previewing}

\item{digits:}{Number of digits for the kable tbale}
}
\description{
Get overview over the data using the skimr packages
}
